requirements.txt
langchain==0.3.7
langchain-google-genai==2.0.7
langchain-community==0.3.7
langchain-core==0.3.19
google-generativeai==0.8.3
chromadb==0.5.20
sentence-transformers==3.3.1
faiss-cpu==1.9.0
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
pydantic==2.10.3
numpy==1.26.4
python-multipart==0.0.17
httpx==0.27.2
